deps: 更新 hnu_query 依赖到 3.0.0 (#58)
Reviewed-on: http://code.qnxg.cn/qnxg/back_2026/pulls/58
Co-committed-by: caiwen <3102733279@qq.com>

+++ src/service/netflow/utils.rs
@@ -20,11 +20,11 @@ pub fn parse_year_month(str: &str) -> Option<NaiveDate> {
 }
 
 /// 将字节转为 GB
-pub fn bytes_to_gb(bytes: f64) -> String {
-    if bytes == 0.0 {
+pub fn bytes_to_gb(bytes: usize) -> String {
+    if bytes == 0 {
         "0 GB".to_string()
     } else {
-        format!("{:.2} GB", bytes / 1024.0 / 1024.0 / 1024.0)
+        format!("{:.2} GB", bytes / 1024 / 1024 / 1024)
     }
 }
